chore: reorg split out access checks
